# backend/swagger.yaml
swagger: "2.0"
openapi: "3.0"
info:
  title: "DeliSOS"
  version: "1.0.0"
# host: "66.70.179.133:8080"
host: "delisos.com"
consumes:
- "application/json"
- "application/x-www-form-urlencoded"
produces:
- "application/json"
schemes:
- "http"
...
paths:
...
  "/user/login":
      post:
        tags:
        - 'user'
        summary: Authenticating user
        parameters: 
          - in: body
            name: user
            description: login user.
            required: true
            schema:
              type: object
              required:
                - email
                - password
                - deviceId
                - deviceType
                - fcmToken
                - latitude
                - longitude
              properties:
                email:
                    type: string
                password:
                    type: string
                deviceId:
                    type: string
                deviceType:
                    type: string
                fcmToken:
                    type: string
                latitude:
                    type: string
                    description: 'The value must be between -90 to 90'
                longitude:
                    type: string
                    description: 'The value must be between -180 to 180'    
        responses:
          "200":
            description: Sucessfully login
            schema:
              $ref: '#/definitions/SuccessMessage'
          "400":
            description: Erorr message.
            schema:
              $ref: '#/definitions/errorMessage'
          "500":
            description: Erorr message.
            schema:
              $ref: '#/definitions/errorMessage'
          "404":
            description: Erorr message.
            schema:
              $ref: '#/definitions/errorMessage'     
...
  "/user/logout":
      get:
        tags:
        - 'user'
        summary: for loggedout from the current session.
        parameters: 
          - name: authorization
            in: header
            description: token for validation.
            required: true
            type: string
        
        responses:
          "200":
            description: ok.
            schema:
              $ref: '#/definitions/SuccessMessage'
          "400":
            description: Erorr message.
            schema:
              $ref: '#/definitions/errorMessage'
          "500":
            description: Erorr message.
            schema:
              $ref: '#/definitions/errorMessage'
...
definitions:
  errorMessage:
    type: "object"
    properties:
      code:
        type: "integer"
      message:
        type: "string"
  SuccessMessage:
    type: "object"
    properties:
      code:
        type: "integer"
      message:
        type: "string"  
...